Split ui package by functionality.
