+++ src/main/java/qubexplorer/ui/issues/FileObjectOpenedListener.java
@@ -1,4 +1,4 @@
-package qubexplorer.ui;
+package qubexplorer.ui.issues;
 
 import org.openide.filesystems.FileObject;
