+++ src/main/java/qubexplorer/ui/summary/SonarRunnerTask.java
@@ -1,4 +1,4 @@
-package qubexplorer.ui;
+package qubexplorer.ui.summary;
 
 import java.awt.event.ActionEvent;
 import java.io.IOException;
@@ -21,6 +21,9 @@ import qubexplorer.runner.SonarRunnerCancelledException;
 import qubexplorer.runner.SonarRunnerProccess;
 import qubexplorer.runner.SonarRunnerResult;
 import qubexplorer.runner.SourcesNotFoundException;
+import qubexplorer.ui.ProjectContext;
+import qubexplorer.ui.SonarIssuesTopComponent;
+import qubexplorer.ui.SonarQubeOptionsPanel;
 import qubexplorer.ui.task.Task;
 import qubexplorer.ui.task.TaskExecutionException;
